Move `PostAdapter` to ui package
