+++ app/src/main/java/info/vividcode/android/app/yuyutumblr/MainActivity.kt
@@ -8,7 +8,6 @@ import com.android.volley.Response
 import com.android.volley.toolbox.ImageLoader
 import com.android.volley.toolbox.ImageLoader.ImageCache
 import com.android.volley.toolbox.JsonObjectRequest
-import com.android.volley.toolbox.NetworkImageView
 import com.android.volley.toolbox.Volley
 
 import android.graphics.Color
@@ -20,11 +19,10 @@ import android.support.v4.widget.SwipeRefreshLayout
 import android.support.v7.widget.LinearLayoutManager
 import android.support.v7.widget.RecyclerView
 import android.util.Log
-import android.view.LayoutInflater
 import android.view.Menu
 import android.view.View
-import android.view.ViewGroup
 import android.widget.Button
+import info.vividcode.android.app.yuyutumblr.ui.PostAdapter
 
 import java.util.ArrayList
 
@@ -62,75 +60,7 @@ class MainActivity : Activity() {
 
     }
 
-    private class Photo(val width: Int, val height: Int, val url: String)
-
-    class PostAdapter// Provide a suitable constructor (depends on the kind of dataset)
-    (private val mImageLoader: ImageLoader) : RecyclerView.Adapter<PostAdapter.ViewHolder>() {
-        private val mList: MutableList<JSONObject>
-
-        val lastItem: JSONObject?
-            get() = if (mList.size == 0) null else mList[mList.size - 1]
-
-        // Provide a reference to the views for each data item
-        // Complex data items may need more than one view per item, and
-        // you provide access to all the views for a data item in a view holder
-        class ViewHolder(// each data item is just a string in this case
-                var mView: View) : RecyclerView.ViewHolder(mView)
-
-        init {
-            mList = ArrayList()
-        }
-
-        // Create new views (invoked by the layout manager)
-        override fun onCreateViewHolder(parent: ViewGroup,
-                                        viewType: Int): PostAdapter.ViewHolder {
-            // create a new view
-            val v = LayoutInflater.from(parent.context)
-                    .inflate(R.layout.post, parent, false)
-            // set the view's size, margins, paddings and layout parameters...
-            return ViewHolder(v)
-        }
-
-        // Replace the contents of a view (invoked by the layout manager)
-        override fun onBindViewHolder(holder: ViewHolder, position: Int) {
-            // - get element from your dataset at this position
-            // - replace the contents of the view with that element
-            try {
-                val post = mList[position]
-                if (post.getString("type") == "photo") {
-                    val p = getAppropriateSizePhotoObject(post.getJSONArray("photos").getJSONObject(0))
-                    //String url = .getJSONObject("original_size").getString("url");
-
-                    val v = holder.mView.findViewById<View>(R.id.image) as NetworkImageView
-                    v.setImageUrl(p!!.url, mImageLoader)
-                    v.minimumHeight = p.height
-                    v.minimumWidth = 36
-                    v.maxHeight = p.height
-                    v.maxWidth = p.width
-
-                    v.setDefaultImageResId(android.R.drawable.ic_menu_rotate)
-                    v.setErrorImageResId(android.R.drawable.ic_delete)
-
-                    //FrameLayout fl = (FrameLayout) convertView.findViewById(R.id.image_container);
-                    //fl.removeAllViews();
-                    //fl.addView(v);
-                }
-            } catch (e: JSONException) {
-                Log.d("error", "error", e)
-            }
-
-        }
-
-        // Return the size of your dataset (invoked by the layout manager)
-        override fun getItemCount(): Int {
-            return mList.size
-        }
-
-        fun add(items: List<JSONObject>) {
-            mList.addAll(items)
-            notifyDataSetChanged()
-        }
-    }
+    class Photo(val width: Int, val height: Int, val url: String)
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -269,7 +199,7 @@ class MainActivity : Activity() {
          * },
          * {"blog_name":"anisample","id":52300690971,"post_url":"http:\/\/anisample.tumblr.com\/post\/52300690971","slug":"","type":"audio","date":"2013-06-06 14:16:32 GMT","timestamp":1370528192,"state":"published","format":"html","reblog_key":"yIm7xUJf","tags":["\u3086\u3086\u5f0f","\u91ce\u3005\u539f\u3086\u305a\u3053","\u5927\u4e45\u4fdd\u7460\u7f8e"],"short_url":"http:\/\/tmblr.co\/Zj_7osmjNJOR","highlighted":[],"note_count":7,"artist":"\u91ce\u3005\u539f \u3086\u305a\u3053","album":"\u3086\u3086\u5f0f","track_name":"\u3053\u306e\u4e0a\u304c\u3063\u305f\u306e\u3069\u3046\u3057\u3088\u3046","album_art":"http:\/\/25.media.tumblr.com\/tumblr_mnz6bkviyS1s85v8so1_1370528192_cover.jpg","caption":"","player":"\u003Cembed type=\u0022application\/x-shockwave-flash\u0022 src=\u0022http:\/\/assets.tumblr.com\/swf\/audio_player.swf?audio_file=http%3A%2F%2Fwww.tumblr.com%2Faudio_file%2Fanisample%2F52300690971%2Ftumblr_mnz6bkviyS1s85v8s&color=FFFFFF\u0022 height=\u002227\u0022 width=\u0022207\u0022 quality=\u0022best\u0022 wmode=\u0022opaque\u0022\u003E\u003C\/embed\u003E","embed":"\u003Ciframe class=\u0022tumblr_audio_player tumblr_audio_player_52300690971\u0022 src=\u0022http:\/\/anisample.tumblr.com\/post\/52300690971\/audio_player_iframe\/anisample\/tumblr_mnz6bkviyS1s85v8s?audio_file=http%3A%2F
          */
-        private fun getAppropriateSizePhotoObject(photoInfo: JSONObject): Photo? {
+        fun getAppropriateSizePhotoObject(photoInfo: JSONObject): Photo? {
             var photo: Photo? = null
             try {
                 val arr = photoInfo.getJSONArray("alt_sizes")
